requirements.txt
snowflake-connector-python==3.12.3
pandas==2.2.3
numpy==1.26.4
matplotlib==3.9.2
seaborn==0.13.2
plotly==5.24.1
scipy==1.14.1
scikit-learn==1.5.2
jupyter==1.1.1
jupyterlab==4.3.3
pyarrow==17.0.0
python-pptx==1.0.2
pyyaml==6.0.2
python-dotenv==1.0.1
tabulate==0.9.0
kaleido==0.2.1
